requirements.txt
boto3
requests
load_dotenv
yt-dlp